chore(compare): rename to ticker board
