+++ ourportfolios/pages/__init__.py
@@ -4,7 +4,7 @@
 from . import landing
 from . import analyze
 from . import ticker_analysis
-from . import compare
+from . import ticker_board
 from . import framework
 from . import select
 from . import industry_analysis
@@ -13,7 +13,7 @@ __all__ = [
     "landing",
     "analyze",
     "ticker_analysis",
-    "compare",
+    "ticker_board",
     "framework",
     "select",
     "industry_analysis",
