+++ ourportfolios/pages/ticker_board/controls.py
