+++ ourportfolios/pages/ticker_board/state.py
