+++ requirements.txt
@@ -1,0 +1,101 @@
+alembic==1.18.4
+annotated-types==0.7.0
+anyio==4.12.1
+apscheduler==3.11.2
+asttokens==3.0.1
+asyncpg==0.31.0
+beautifulsoup4==4.14.3
+bidict==0.23.1
+certifi==2026.1.4
+cffi==2.0.0
+charset-normalizer==3.4.4
+click==8.3.1
+comm==0.2.3
+contourpy==1.3.3
+cycler==0.12.1
+debugpy==1.8.20
+decorator==5.2.1
+dotenv==0.9.9
+et-xmlfile==2.0.0
+executing==2.2.1
+fonttools==4.61.1
+granian==2.7.1
+greenlet==3.3.2
+h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
+idna==3.11
+importlib-metadata==8.7.1
+ipykernel==7.2.0
+ipython==9.10.0
+ipython-pygments-lexers==1.1.1
+jedi==0.19.2
+jupyter-client==8.8.0
+jupyter-core==5.9.1
+kiwisolver==1.4.9
+mako==1.3.10
+markdown-it-py==4.0.0
+markupsafe==3.0.3
+matplotlib==3.10.8
+matplotlib-inline==0.2.1
+mdurl==0.1.2
+nest-asyncio==1.6.0
+numpy==2.4.2
+openpyxl==3.1.5
+-e file:///home/dank/Documents/Codebases/ourportfolios-net/ourportfolios
+-e file:///home/dank/Documents/Codebases/ourportfolios-net/ourscheduler
+packaging==26.0
+pandas==3.0.1
+parso==0.8.6
+pexpect==4.9.0
+pillow==12.1.1
+platformdirs==4.9.2
+prompt-toolkit==3.0.52
+psutil==7.2.2
+psycopg2==2.9.11
+psycopg2-binary==2.9.11
+ptyprocess==0.7.0
+pure-eval==0.2.3
+pycparser==3.0
+pydantic==2.12.5
+pydantic-core==2.41.5
+pygments==2.19.2
+pyparsing==3.3.2
+python-dateutil==2.9.0.post0
+python-dotenv==1.2.1
+python-engineio==4.13.1
+python-multipart==0.0.22
+python-socketio==5.16.1
+pytz==2025.2
+pyzmq==27.1.0
+redis==7.2.0
+reflex==0.8.27
+reflex-hosting-cli==0.1.61
+requests==2.32.5
+rich==14.3.3
+ruff==0.15.2
+seaborn==0.13.2
+simple-websocket==1.1.0
+six==1.17.0
+soupsieve==2.8.3
+sqlalchemy==2.0.46
+sqlmodel==0.0.37
+squarify==0.4.4
+stack-data==0.6.3
+starlette==0.52.1
+tenacity==9.1.4
+tornado==6.5.4
+traitlets==5.14.3
+typing-extensions==4.15.0
+typing-inspection==0.4.2
+tzlocal==5.3.1
+urllib3==2.6.3
+vnai==2.3.9
+vnstock==3.4.2
+vnstock-ezchart==0.0.3
+watchfiles==1.1.1
+wcwidth==0.6.0
+wordcloud==1.9.6
+wrapt==2.1.1
+wsproto==1.3.2
+zipp==3.23.0
